[feat/fix] Add ability to detect network reachability on FireOS devices
Looks like FireOS does not correctly report capabilities of the default network correctly, reporting a validated state when the internet is not actually reachable. After much experimentation on our project we were able to nail down the actions and broadcasts the OS uses to perform connectivity checks. While these are undocumented by Amazon, they have been in wide use for a while and have been presented as a side note in some presentation slides from Amazon.

This implementation will detect if the device in use is a FireOS device and only enable the extra checks if so. The data received from FireOS will override the mIsInternetReachable in ConnectivityReceiver instead of replacing it, allowing for other overrides to be done in a similar way later on for other platforms that do not use NetworkCapabilities correctly.