feat: Add reachability request timeout to improve handling of bad network connections (#302 by @davidstritzl)
This fixes an issue where a reachability check would not produce a result until a response from the reachability request is received. The fetch API, which is used for this purpose, does not have a built-in timeout mechanism, therefore requests can, in theory, stay pending forever. In case of a bad connection with packet loss, this can mean that no reachability issues are detected, even though there is effectively no internet connection. This commit aims to fix this issue by implementing a user-configurable request timeout.