fix(ios): avoid memory leak from ssid APIs by adding explicit config (#560)
* Fix #420 - Location Permissions Required

`CNCopyCurrentNetworkInfo` will leak without location permissions being enabled.
Added comments for depreciations & implementation notes that need future work.

* 420 - Pass config to native iOS & gate on new prop

Added new exported method for passing configuration to native iOS.  Storing it there on the instance as a property and gating based on that config.  It will not fetch SSIDs if config is nil.  Also cleaned up a bit of typing.

* 420 - Fix mock
* 420 Updated README

Co-authored-by: Mike Hardy <github@mikehardy.net>