fix(windows): fix crash in getIpAddressSync
Even if not properly documented, some properties of
`windows.networking.connectivity.connectionprofile` can throw.
It's the case of `NetworkAdapter` since it's then calling
`check_hresult` which will throw if the returned hresult is an error one.

Since `RNCNetInfo::getIpAddressSync` is marked as `noexcept`, any exception
thrown in this method will call `std::terminate`.

This commit is to return `"unknown"` in this case.